requirements.txt
streamlit
fastapi
uvicorn
moviepy
pillow
numpy
matplotlib
pysrt
python-dotenv
groq
torch
transformers
accelerate
peft
